fix: allow projects to use database
For context, encore internally uses a proxy to the db to make requests. Also, encore does static analysis to figure out what services should be allowed access to which databases (check https://discord.com/channels/814482502336905216/814482502336905219/1364366338788823191). So, the workaround for now is to declare encore.service.ts for every service and declare the Kysely instance somewhere inside the service folder